created new component for ListofUnits
